backend/app/services/aider_integration.py
import asyncio
from typing import Optional
import aider.main as aider

class AiderService:
    def __init__(self):
        self.sessions = {}

    async def process_message(self, message: str, project_id: int) -> str:
        if project_id not in self.sessions:
            self.sessions[project_id] = await self._create_aider_session(project_id)
        
        session = self.sessions[project_id]
        response = await self._run_aider_command(session, message)
        return response

    async def _create_aider_session(self, project_id: int):
        # Initialize Aider session with project-specific configuration
        # This is a placeholder for actual Aider initialization
        return {
            "project_id": project_id,
            "context": []
        }

    async def _run_aider_command(self, session: dict, message: str) -> str:
        # This is where we'll integrate with the actual Aider CLI
        # For now, return a placeholder response
        return f"Aider processed: {message}"