backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.27
pydantic==2.6.1
pydantic-settings==2.1.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.9
psycopg2-binary==2.9.9
redis==5.0.1
websockets==12.0
python-dotenv==1.0.1
bcrypt==4.1.2
aider-chat==0.18.2